==> step2_data_cleaning.ipynb <==
import pandas as pd

# Load Dataset

df = pd.read_csv("final_data/initial_dataset.csv") 
df.head()

# Check Data Types

print(df.dtypes)

# Remove Irrelevant Columns

Columns `pics`, `resp`, `url`, `Unnamed: 0` are no longer relevant in the dataset as we would not be using these fields in the analysis downstream.

As such, we would remove these columns.

df = df.drop(columns=["pics", "resp", "url", "Unnamed: 0", "fake"])

# Rename Columns

We will rename columns to a more descriptive name so that it is more intuitive.

`name_x` will be renamed to `user_name`

`name_y` will be renamed to `business_name`

`text` will be renamed to `review`

`category` will be renamed to `business_desc`

`policy` will be renamed to `label`

df.rename(
    columns={
        "name_x": "user_name",
        "name_y": "business_name",
        "text": "review",
        "category": "business_desc",
        "policy": "label"
    },
    inplace=True
)

df.head()

# Convert Data Type

The `time` column is provided in Unix time (milliseconds). 

Since this format is not easily readable or intuitive, we will convert it to a standard datetime format for easier interpretation and analysis.

df['time'] = pd.to_datetime(df['time'], unit='ms')

# Check Null Values in each column

print(df.isnull().sum())  

# Handling Missing Values

Although columns like `user_id`, `user_name`, and `business_name` contain missing values, they will not serve as key features in our model development. 

Therefore, we will not perform any imputation or removal for these columns.

### `rating`

The `rating` column is a numerical variable. There are 74 rows with missing values, which we will impute using the median of the available ratings.

Since the number of missing rows is small, median imputation is unlikely to significantly alter the overall distribution of ratings.

median_rating = df['rating'].median()
df['rating'] = df['rating'].fillna(median_rating)

### `business_desc`

There are 5 rows with missing `business_desc` values. Since `business_desc` is a textual categorical column and will be important for downstream analysis, we will drop these 5 rows. 

They represent a very small proportion of the dataset, so their removal will not significantly affect our analysis.

df = df.dropna(subset=['business_desc'])

# After Handling of Missing Values

print(df.isnull().sum())  

All missing values for the relevant columns to be used in our model development has been handled.

# Set Unique ID

df.insert(0, "review_id", range(1, len(df) + 1))

# Save Dataset

df.to_csv("final_data/cleaned_dataset.csv", index=False, encoding='utf-8-sig')
